refactor(app): extract useWarpLaunch hook (92y)
Moves warp state (pendingWarpLabelName/Target/VariableDrafts, warpTempFilePathRef)
and all warp callbacks (cleanupWarpTempFile, resetWarpLaunchState,
handleConfirmWarpVariables, handleWarpToLabel) out of App.tsx into
src/hooks/useWarpLaunch.ts. cleanupWarpTempFile is returned by the hook
and threaded into useGameExecution.
